(maint) Reset Puppet::Parser::Functions early in TestHelper.initialize
External tests which rely on being able to simply require 'puppet' and
then load a function for test relied on Puppet::Parser::Functions.reset
being called during the class loading.  Now that that has moved into
Puppet initialization, we simulate this in TestHelper by calling it
before specs are loaded.