(PUP-1438) yumrepos type does not handle white space in repo files
The yum repository definition files from Redhat come with space between
the key and the value (such as "enabled = 1"). The Redhat setting for
the sslVerify key is also a "1" even though the man page lists acceptable
values as (True|False)

This commit modifies the regex for the yumrepo type to allow puppet to
parse the Yum repo config files from Redhat.