(maint) Don't duplicate hash keys in rspec tests
RSpec examples can be restricted by using `:if`: and `:unless` values in
the example metadata. However these are hash keys, and defining two
`:if` statements will cause the latter definition to overwrite the
first. This commit resolves this issue by defining a single expression
evaluated by `:if` instead of having two separate clauses.