(maint) Unify to_pson in FormatSupport module, just like with to_msgpack
This removes all unnecessary to_pson definitions and creates a common
one in the Puppet::Network::FormatSupport module instead just like with
to_msgpack.

There are still some to_pson definitions left but that is because they
also add a document type structure. That cannot be removed until the
next major version as it would change the on wire format.