(PUP-576) Update store_configs test for Debian
The store_configs/enc_provides_node_when_storeconfigs_enabled.rb test
has been updated to use Activerecord 3.2.x, but this version of
Activerecord no longer supports the sqlite3-ruby gem package available
on Debian via apt.  So we are installing the newer sqlite3 gem.  As an
added wrinkle, this is the same name as the core sqlite3 package on
Debian, running us afoul of PUP-1073 (duplicate names for packages
handled by different providers).  We are working around that with a
separate puppet apply for the sqlite3 installation.