(PUP-530) Make install.rb check for hiera before installing
From ticket:
The `install.rb` script for puppet checks for facter and will refuse to install
puppet if facter is not on the system. However, puppet also cannot function
correctly without hiera and the `install.rb` script does not check for that.
The installation script that we ship in the tarball should also check for hiera
being present.

This commit adds hiera to the list of prerequisite libraries required before
install.rb will install puppet.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>