(PUP-753) Benchmark puppet compiling w/many modules
This sets up a scenario in which we are benchmarking the puppet compiler
with a large number of modules, but very simple manifests. The generated
modules have 2 manifests each an public and an internal. The site.pp for
the environment simply includes all of the modules' public classes. The
public classes in turn include their own internal class. The internal
class then places a single "notify" resource in the catalog.

This also ends up laying down a general structure for benchmark
scenarios. Each scenario has support files under `bencmarks/<name of
scenario>` and has a rake namespace under the `benchmark` namespace.
Each namespace should have at least tasks for `generate` and `run`.