(maint) Pass Facter.search arrays instead of strings
The implementation of Facter.search expects `*args` while Puppet was
defaulting to a colon separated string. This isn't a valid path but
until Facter commit 6c6c394 warnings were almost always squelched which
prevented this from being exposed. This commit ensures that we split up
our (semi)colon delimited paths into an array and passes them to Facter
in the manner that it expects.