(PUP-753) Extract benchmark impl into file
The code for setting up and running the benchmark is very specific to
this particular scenario and also too large to keep in a rake file. This
extracts it out into a file that lives inside the scenario. This should
turn into a common interface that all benchmark scenarios can implement
and should allow the maintenance of scenarios to remain separated and
the creation of new ones to be vastly simplified.