(PUP-753) Generate scenario tasks from filesystem
Since the scenario execution is just defined by a ruby class that lives
in the scenario directory, the scenario tasks can now be generated. This
extracts the description of a scenario into a file (`description`) and
generates a set of rake tasks for each scenario that is in the
benchmarks directory.

This also removes the top-level benchmark task because right now all
scenarios will be using the same class name (Benchmarker), which would
cause name collisions when there are multiple scenarios being run in the
process.