(PUP-753) Constrain ruby-prof solely to the development group
Excludes ruby-prof from the test group so that we can run specs on test
machines without requiring it.  In particular, ruby-prof makes running
specs on Windows difficult unless devkit is present or a ruby installation
for windows has been pre-built with the ruby-prof gem in place.  We may
need to revisit this when we get a ci setup for benchmarking.