(PUP-1407) Verify that CRL times are within expected ranges
Modifies the test so that we verify the CRL last and next update times are
in UTC time (per the RFC) and that the number of seconds since the epoch
are within 5 minutes (plus or minus) of the expected value.

Also modifies the test to use public APIs to access the CA cert and
private key.