(PUP-753) Benchmark system startup
This adds a simple benchmark to check how long it takes to get puppet to
respond to commands on the CLI. This is an important aspect of percieved
performance for many users. And for anyone trying to script puppet
commands, this becomes a *very* important aspect of actual system
performance.