(PUP-1407) Verify extensions are still present after revocation
Previously, we were only testing that the crlNumber and authorityKeyIdentifier
extensions were present when a CRL is generated, but not when it is revoked.

This adds helpers for testing the extensions and updates both cases to use them.
It also is more explicit in verifying that the authorityKeyIdentifier extension
has the same value as the subjectKeyIdentifier from the issuing CA cert.