(PUP-1407) Fix specs on 187
In ruby 1.8.7, OpenSSL::X509::Name does not include Comparable, so `==` is
not based on the `<=>` operator. As a result, comparing two Name objects
that have the same string representation will return false. In ruby 1.9,
the class includes the Comparable module. Changed the test to just compare
the String representations.

The other test was failing for me on ruby 187, and it seemed completely
unnecessary to compute 2**158 every time the test runs.