(maint) Hold token name in local variabel
The tokens name was fetch multiple times in quick succession. The lexer
is in a very tight inner loop and so any reduction of method calls can
produce improvements in performance. Although most work for language
performance is focussed on the new parser, this is a simple, small win
that we can easily take now.