(PUP-1585) cron: fix resource duplication in crontab provider
The regression fix for #19876 caused yet another regression in cron
resources that manage the target property, but not the user property.
The earlier fix made sure that a cron resource would not try and manage
a matching record from another user's crontab. It did so by comparing
the record's target to the resource's user property.

When the user is unmanaged, but the target is specified instead, this
would cause puppet to ignore the record on disk and add it once more.

This is fixed by comparing the record's target to either the user
property or the target property. If there is a should-value for the
user, it is used, otherwise the provider falls back to the target
should-value.