(maint) Fork for each scenario
Because of the way puppet initializes, we can't safely re-initialize in
the same process. In order to run all of the scenarios we need to make
sure there is a new process for each one. This changes from having the
"all" tasks re-run rake for each of these instead of just making them
dependencies of the "all" task.