(PUP-1619) Fix problems with Ruby 1.8.7 Range min/max & Infinity
Billions of blue blistering boiled and barbecued barnacles, the
Ruby 1.8.7 implementation of Range min/max is based on Enumeration.
This does not work well when the range has +/- Infinity at one of
the ends. There is probably a "break" when it reaches Infinity...

*sign*

This commit changes the implementation from using a Range to using
an Array instead of a Range.