(PUP-1661) Add Skeleton Face to Module command
From discussions with some of the Forge team, there was idea that it would be a nice feature to have a command within the module code that would generate a module from a given 'skeleton': a pre-made directory of files to use as a base for a new module.  

This is the first step towards that larger implementation. For now the module command will have a new face 'skeleton' which, when run. will return the module default path and the module skeleton directory. In the future, you'll be able to provide a local or remote directory to the command (for example, @garethr's excellent `puppet-module-skeleton` directory, which contains nice-to-have extras such as a Guardfile, .travis.yml file and the like) which will then be used by the `puppet module generate` command as a base.