(maint) Fix pkgin package provider spec to ensure resource uses the pkgin provider.
GH-2023 included a change that added a version onto the resource.  However, since
a provider was not declared, it used the system's default provider.

Because of this, running the spec on a system with a default provider that doesn't
supoprt the versionable feature would cause the spec to fail.

The fix is to set the provider on the resource and move the ensured version to the
particular test case.