(PUP-1584) module_tool uses current_environment with target_dir overrides
The module tool was relying on the implicit creation and caching of
legacy environments based on puppet settings for environment and
modulepath.  But this does not work for the new directory style of
environments.  So we are being explicit about using the current
environment from the context, and overriding modulepath to prepend a
target_dir if given, and then directly inserting the environment to be
used into the options rather than expecting the module_tool applications
to retrieve the correct environment from any other loaders.

Testing is now simpler, but because we are now comparing different but
equal environments, I added equality and hash to node/environment to
facilitate this.

This in turn broke tests, because of
Puppet::Node::Environment#known_resource_types which caches using the
global $known_resource_types.  Previously this would always clear itself
when called if $known_resource_types.environment
!= self, but with the new node equality, state was now leaking between
tests.  The test_helper now sets $known_resource_types nil after each
test.