(PUP-1707) Fix possible gsubs on nil in man page ERB template
In several places, this template would attempt to fall back to summaries if long
descriptions were absent. However, before deciding to use the summary, it would
attempt to do a gsub on nil, which would blow up the entire process.

This commit does the fallback earlier, and adds final fallback text so we won't
be using text methods on nil if the summary is also absent.