(Maint) Ensure there is a fully ruby env
The tests were neglecting to require 'rubygems', which meant that they
were not working with the full environment of modules that puppet needs
to run. Setting up the environment is the responsibility of the code
that is going to load puppet, which is why this change was made here,
rather than changing puppet to require rubygems itself.