(PUP-30) Add SubLocatableExpression to handle "embedded" expressions
This expression helps with handling areas of the source that requires
douple/separate lexing (heredoc, etc). It establishes a context
from which its children can resolve their offsets.

This is requires to make the offset calculations fast (no need to
relocate them). Instead the relative location is captured in a
SubLocateableExpression.