(PUP-1678) List environment settings instead of modules
In talking with consumers of the environment information it turns out
that listing the modules isn't the desired feature. Instead they really
want the settings of the environment (modulepath, manifest). Other
settings may be added later.

This has the added benefit that listing the environments is now a much
cheaper operations since it doesn't need to load the module data.